requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
alembic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]  
pydantic-settings 
python-multipart